[CMD] Attempt to fix app launching

+++ kernel/core/invoke.c
@@ -39,12 +39,12 @@ int KeInvokeMainArgs(void *entry, const char *image_path, const char *command_li
     return ((ENTRY)entry)(argc, argv);
 }
 
-int KeInvokeWMain(void *entry, const char *image_path, const uint32_t *command_line,
+int KeInvokeWMain(void *entry, const char *image_path, const uint16_t *command_line,
                   void *stack, uint32_t stack_size) {
-    typedef int (*ENTRY)(int, uint32_t **);
-    uint32_t buffer[256];
-    uint32_t *argv[32];
-    uint32_t image[256];
+    typedef int (*ENTRY)(int, uint16_t **);
+    uint16_t buffer[256];
+    uint16_t *argv[32];
+    uint16_t image[256];
     int argc = 1, i = 0, j;
     (void)stack;
     (void)stack_size;
@@ -59,7 +59,7 @@ int KeInvokeWMain(void *entry, const char *image_path, const uint32_t *command_l
         while (buffer[i] == ' ' || buffer[i] == '\t') i++;
         if (!buffer[i]) break;
         if (buffer[i] == '"' || buffer[i] == '\'') {
-            uint32_t quote = buffer[i++];
+            uint16_t quote = buffer[i++];
             argv[argc++] = &buffer[i];
             while (buffer[i] && buffer[i] != quote) i++;
         } else {
