+++ kernel/core/invoke.h
@@ -4,7 +4,7 @@
 int KeInvokeMain(void *entry, void *stack, uint32_t stack_size);
 int KeInvokeMainArgs(void *entry, const char *image_path, const char *command_line,
                      void *stack, uint32_t stack_size);
-int KeInvokeWMain(void *entry, const char *image_path, const uint32_t *command_line,
+int KeInvokeWMain(void *entry, const char *image_path, const uint16_t *command_line,
                   void *stack, uint32_t stack_size);
 int KeInvokeWinMain(void *entry, void *image, const char *command_line,
                     void *stack, uint32_t stack_size);
